+++ CGALWrapper/PolygonBoolean2_EEK.cpp
@@ -52,68 +52,68 @@ int PolygonBoolean2_EEK_Intersect_P_P(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Intersect_P_P(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Intersect_P_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Intersect_P_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Intersect_PWH_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Intersect_PWH_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Difference_P_P(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Difference_P_P(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Difference_P_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Difference_P_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Difference_PWH_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Difference_PWH_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_SymmetricDifference_P_P(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::SymmetricDifference_P_P(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_SymmetricDifference_P_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::SymmetricDifference_P_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_SymmetricDifference_PWH_PWH(void* ptr1, void* ptr2)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::SymmetricDifference_PWH_PWH(ptr1, ptr2, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
 
 int PolygonBoolean2_EEK_Complement_PWH(void* ptr)
 {
 	buffer.clear();
 	PolygonBoolean2<EEK>::Complement_PWH(ptr, buffer);
-	return buffer.size();
+	return (int)buffer.size();
 }
